# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inazuma API",
        "description": "API for Inazuma Eleven",
        "contact": {
            "email": "gomez.sergio.smr@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/stats": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "List all player stats",
                "operationId": "a50eb4df8ec353c551c12ee48a36b301",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "playerId[eq]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "version[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ie1"
                        }
                    },
                    {
                        "name": "GP[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "TP[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "kick[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "body[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "control[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "guard[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "speed[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "stamina[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "guts[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    },
                    {
                        "name": "freedom[gt]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 80
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Stats"
                ],
                "summary": "Create new player stats",
                "operationId": "1f4abc9eace146bb71aa17e3e83e51be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stats created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/stats/{id}": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get specific player stats",
                "operationId": "4e7f5670d21ae4dafdcd08f6d23ebd03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Stats not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stats"
                ],
                "summary": "Update player stats",
                "operationId": "e8dcabab665c861f07a3d4d952ba3786",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stats updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stats"
                ],
                "summary": "Delete player stats",
                "operationId": "c1c01b048fa27f76b04f72d93d2e16a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stat deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StatRequest": {
                "required": [
                    "playerId",
                    "version",
                    "GP",
                    "TP",
                    "kick",
                    "body",
                    "control",
                    "guard",
                    "speed",
                    "stamina",
                    "guts",
                    "freedom"
                ],
                "properties": {
                    "playerId": {
                        "type": "integer",
                        "example": 1
                    },
                    "version": {
                        "type": "string",
                        "example": "ie1"
                    },
                    "GP": {
                        "type": "integer",
                        "example": 50
                    },
                    "TP": {
                        "type": "integer",
                        "example": 100
                    },
                    "kick": {
                        "type": "integer",
                        "example": 80
                    },
                    "body": {
                        "type": "integer",
                        "example": 70
                    },
                    "control": {
                        "type": "integer",
                        "example": 75
                    },
                    "guard": {
                        "type": "integer",
                        "example": 65
                    },
                    "speed": {
                        "type": "integer",
                        "example": 90
                    },
                    "stamina": {
                        "type": "integer",
                        "example": 85
                    },
                    "guts": {
                        "type": "integer",
                        "example": 95
                    },
                    "freedom": {
                        "type": "integer",
                        "example": 60
                    }
                },
                "type": "object"
            },
            "StatResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Stat"
                    }
                },
                "type": "object"
            },
            "Stat": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "player_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "version": {
                        "type": "string",
                        "example": "ie1"
                    },
                    "GP": {
                        "type": "integer",
                        "example": 50
                    },
                    "TP": {
                        "type": "integer",
                        "example": 100
                    },
                    "kick": {
                        "type": "integer",
                        "example": 80
                    },
                    "body": {
                        "type": "integer",
                        "example": 70
                    },
                    "control": {
                        "type": "integer",
                        "example": 75
                    },
                    "guard": {
                        "type": "integer",
                        "example": 65
                    },
                    "speed": {
                        "type": "integer",
                        "example": 90
                    },
                    "stamina": {
                        "type": "integer",
                        "example": 85
                    },
                    "guts": {
                        "type": "integer",
                        "example": 95
                    },
                    "freedom": {
                        "type": "integer",
                        "example": 60
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StatCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Stat"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/stats?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/stats?page=3"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/stats?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 35
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}